Refactor main package organization.
Reorganized the components in the main package into a subpackage to make
sure they are at the same level as the rest of the components.
