+++ src/main/java/nl/tudelft/pixelperfect/event/EventLog.java
@@ -2,8 +2,8 @@ package nl.tudelft.pixelperfect.event;
 
 import com.jme3.network.Server;
 
-import nl.tudelft.pixelperfect.Spaceship;
 import nl.tudelft.pixelperfect.client.EventsMessage;
+import nl.tudelft.pixelperfect.game.Spaceship;
 
 import java.util.ArrayList;
