+++ src/main/java/nl/tudelft/pixelperfect/game/Game.java
@@ -1,4 +1,4 @@
-package nl.tudelft.pixelperfect;
+package nl.tudelft.pixelperfect.game;
 
 import java.io.IOException;
