+++ src/main/java/nl/tudelft/pixelperfect/gamestates/StartState.java
@@ -1,7 +1,7 @@
 package nl.tudelft.pixelperfect.gamestates;
 
 
-import nl.tudelft.pixelperfect.Game;
+import nl.tudelft.pixelperfect.game.Game;
 
 /**
  * First state that the game loads.
