+++ src/main/java/nl/tudelft/pixelperfect/gamestates/WonState.java
@@ -1,6 +1,6 @@
 package nl.tudelft.pixelperfect.gamestates;
 
-import nl.tudelft.pixelperfect.Game;
+import nl.tudelft.pixelperfect.game.Game;
 
 /**
  * State for when you have won the game.
