+++ src/test/java/nl/tudelft/pixelperfect/event/EventLogTest.java
@@ -17,10 +17,10 @@ import org.junit.Test;
 
 import com.jme3.network.Server;
 
-import nl.tudelft.pixelperfect.Spaceship;
 import nl.tudelft.pixelperfect.client.EventsMessage;
 import nl.tudelft.pixelperfect.event.type.AsteroidImpactEvent;
 import nl.tudelft.pixelperfect.event.type.FireOutbreakEvent;
+import nl.tudelft.pixelperfect.game.Spaceship;
 
 /**
  * Test Suite for the EventLog class.
