+++ src/test/java/nl/tudelft/pixelperfect/gamestates/PlayStateTest.java
@@ -1,9 +1,10 @@
 package nl.tudelft.pixelperfect.gamestates;
 
 import com.jme3.scene.Spatial;
-import nl.tudelft.pixelperfect.Game;
-import nl.tudelft.pixelperfect.Spaceship;
+
 import nl.tudelft.pixelperfect.event.EventScheduler;
+import nl.tudelft.pixelperfect.game.Game;
+import nl.tudelft.pixelperfect.game.Spaceship;
 import nl.tudelft.pixelperfect.gui.GameHeadsUpDisplay;
 import org.junit.Before;
 import org.junit.Test;
