+++ src/test/java/nl/tudelft/pixelperfect/player/GunnerTest.java
@@ -4,10 +4,11 @@ import static org.hamcrest.CoreMatchers.equalTo;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertThat;
-import nl.tudelft.pixelperfect.Spaceship;
 
 import org.junit.Before;
 import org.junit.Test;
+
+import nl.tudelft.pixelperfect.game.Spaceship;
 
 /**
  * Test suite for the Gunner class.
